requirements.txt
Flask==3.0.2
pandas==2.2.1
openpyxl==3.1.2
qrcode[pil]==7.4.2
Pillow==10.3.0
python-pptx==0.6.23
fpdf==1.7.2
gunicorn==21.2.0
numpy==1.26.4
scipy==1.13.0
python-docx==1.1.0
docxtpl==0.16.7
docxcompose==1.4.0
groq
